adding basic slack communication functionality
Be sure to take a look the config/settings.php as well as env.local.php
Not sure if it is just me or not but it appears that arrays do not really work in the env file, so fetching them into the settings file does not want to work.  For now, just be sure to set the settings.php file up in the config/local path to configure where the slack notifications should go.  I set up as an array for the isntance of notifying multiple places/accounts.